Bugfixes to create valid JPX files
I modified a few small glitches so that SIPI is producing valid JPEG2000 files which run against the validator Jpylyzer. The only "issue" is that SIPI is producing a "jpx" file, thus the "brand" is "jpx" and not "jp2". This can easily be fixed in the validator (I will make an issue there).