BROKEN PIPE
Possible bugfix: if a socket has been closed by peer while writing to it, a SIGPIPE signal is sent to the process which terminates the process (that is sipi). SIGPIPE signals are now catched and – besides writing a log message – ignored. This sipi should handle broken pipes the peer has closed gracefully. Please test!