Colorprofile (#157)
* Bugfix for gray value j2k

* Conformance to J2K test files

Added conformance to read all but one of the J2K reference files. One of the file still poses a problem with a strange kakadu error:
file2.jp2: Attempting to open a tile which has already been discarded or permanently closed!

* Bugfix for gray value j2k

* Conformance to J2K test files

Added conformance to read all but one of the J2K reference files. One of the file still poses a problem with a strange kakadu error:
file2.jp2: Attempting to open a tile which has already been discarded or permanently closed!

* rebase conficts

* Removed debugginh output

* Added ICC profiles

* ongoing…

* Fixed memroy leaks