Improvements (#170)
* Bugfix with tag EXIFTAG_IMAGEUNIQUEID

the EXIFTAG_IMAGEUNIQUEID is a ‚\0’-terminated string and not a char-pointer with length 33 - even if it always has the length 33…

* Some improvements

- error message if script init file is not found
- —query (-x) option for getting all metadata of a file

* Adding subdirs to image directory

* Corrected some errors

* ongoing…

* More work...

* Searching for ICC problem with gray ICC profile

* Fixed bug with ICC profile (crash)

* Subdirs...

* Subdirs works

* Extended documentation of special Lua functions

* Changed config „docroute“ to „wwwroute“

Since the config variables „docroot“ and „docroute“ are easy to mix, „docroute“ has bee changed to „wwwroute“

* If „-f“ („—format“) missing, file-extension is used

if the format specificer „-f“ or „—format“ is missing, SIPI tries to determine the output file format using the extension of the filename. Supported are „jpx“, „jp2“, „tif“, „tiff“, „jpg“, „jpeg“ and „png“.

* Removed restriction for XMP

* Small Comment fix

* Adaption to different dirent implementations

* Bugfix when using prefix and subdirs

* missing file format -> error message

If the file format cannot be determined (no --format option or missing or unknown file extension), an error message is emitted and the command line conversion terminates.