Added page support for pdf's
PDF files can no ne rendered with an extension of the IIIF syntax:
the image id may be extended by "@" and a pagenumber, e.g. "mypdf.pdf@3" will render only page 3 of the PDF file. In this case, region, size and rotation parameters do work and have to be supplied!